refactor: rename TocNode component to TableOfContentsNode and remove unused TocNode.vue
- Updated import in TableOfContents.vue to use the new TableOfContentsNode.vue component.
- Deleted the obsolete TocNode.vue file to streamline the codebase.

This refactor improves clarity and maintainability by consolidating the table of contents component structure.
